:truck: Move page components into components directory
