+++ src/components/README.md
@@ -2,5 +2,7 @@
 
 This directory contains the components that **will be shared by the whole application**.
 
-🖐 **Components** that are used on a single page, should not be placed here.
-They should be located on the specific page inside a `components/` folder ⚠️.
+🖐 **Components** that are used on a single page, should be placed into his
+specific folder under the `pages/` directory.
+
+Components that are used on multiple places should be placed into the `shared/` directory.
