+++ src/components/pages/index/Header/__tests__/header.spec.js
