+++ src/components/pages/index/OpenSource/__tests__/stubs.js
