+++ src/components/pages/index/SectionTitle/__tests__/sectionTitle.spec.js
