+++ src/components/pages/index/SectionTitle/__tests__/stubs.js
