+++ src/components/pages/index/Writings/__tests__/stubs.js
