+++ src/components/pages/index/Writings/index.js
@@ -2,8 +2,8 @@
 import React from 'react'
 
 import { type Post } from 'src/utils/api/ghost/mutators'
-import Wrapper from 'src/components/Wrapper'
-import BlogPost from 'src/components/BlogPost'
+import Wrapper from 'src/components/shared/Wrapper'
+import BlogPost from 'src/components/shared/BlogPost'
 import SectionTitle from '../SectionTitle'
 import styles from './styles.module.css'
