+++ src/pages/index/index.js
@@ -3,12 +3,12 @@ import React from 'react'
 
 import { fetchPosts } from 'src/utils/api/ghost'
 import { fetchRepositories } from 'src/utils/api/github'
-import { type Post } from 'src/utils/api/ghost/mutators'
-import { type Repository } from 'src/utils/github/mutators'
-import Header from './components/Header'
-import Biography from './components/Biography'
-import Writings from './components/Writings'
-import OpenSource from './components/OpenSource'
+import { type Post } from 'src/utils/api/blog/mutators'
+import { type Repository } from 'src/utils/api/github/mutators'
+import Header from 'src/components/pages/index/Header'
+import Biography from 'src/components/pages/index/Biography'
+import Writings from 'src/components/pages/index/Writings'
+import OpenSource from 'src/components/pages/index/OpenSource'
 
 type Props = {
   posts: Array<Post>,
